Fix Bucket Sort.

+++ algostruct/include/sorts/bucket_sort.h
@@ -10,13 +10,13 @@ void BucketSort(Iterator begin, Iterator end,
                 ValueType lower_bound, ValueType upper_bound/* (!)non-include */,
                 const Comparator& comparator) {
     auto count_buckets = static_cast<size_t>(std::distance(begin, end));
-    std::vector<std::vector<ValueType>> buckets(count_buckets + 1);
+    std::vector<std::vector<ValueType>> buckets(static_cast<size_t>(count_buckets) + 1);
     Iterator it = begin;
-    double percent_for_bucket = ((upper_bound - lower_bound) / static_cast<double>(count_buckets))
-            / (upper_bound - lower_bound);
+    double percent_for_bucket = ((static_cast<int64_t>(upper_bound) - lower_bound) / static_cast<double>(count_buckets))
+            / (static_cast<int64_t>(upper_bound) - lower_bound);
     while (it != end) {
         double percent_for_value = (static_cast<double>(*it) - lower_bound)
-                                   / (upper_bound - lower_bound);
+                                   / (static_cast<int64_t>(upper_bound) - lower_bound);
         auto number_bucket_for_value = static_cast<size_t>(percent_for_value / percent_for_bucket);
         buckets[number_bucket_for_value].emplace_back(*it);
         ++it;
